add map util unit tests

+++ src/lib/iface.h
@@ -23,9 +23,9 @@ int mc_open(mc_session * session,
             const enum mc_iface_type iface, const pid_t pid);
 int mc_close(mc_session * session);
 int mc_update_map(const mc_session * session, mc_vm_map * vm_map);
-ssize_t mc_read(const mc_session * session, const uintptr_t addr, 
-                cm_byte * buf, const size_t buf_sz);
-ssize_t mc_write(const mc_session * session, const uintptr_t addr, 
-                 const cm_byte * buf, const size_t buf_sz);
+int mc_read(const mc_session * session, const uintptr_t addr, 
+            cm_byte * buf, const size_t buf_sz);
+int mc_write(const mc_session * session, const uintptr_t addr, 
+             const cm_byte * buf, const size_t buf_sz);
 
 #endif
